Separate training functions and interleave
